First Commmit
Changes compared to the original DomotigaApp:

Changes the homescreen (not done yet). It now shows your locations and when you choice one you will get all the devices from that location.

When reopening the app it will automatically refresh the screen (if you are on a screen with devices) next commit the homescreen wil have this also.

Color scheme has changed from black to white.

All comments/suggestions can be send to cvanbrummelen@merlinsystems.nl